@@ -1,41 +1,36 @@
 ﻿using System.Text;
 using System.Text.RegularExpressions;
+using Crip.AspNetCore.Logging.LongJsonContent.Configuration;
+using Crip.AspNetCore.Logging.LongJsonContent.Services;
 using Crip.Extensions.Tests;
 using Microsoft.Extensions.Options;
 
-namespace Crip.AspNetCore.Logging.Tests.Middlewares;
+namespace Crip.AspNetCore.Logging.LongJsonContent.Tests;
 
-public class LongJsonContentMiddlewareTests
+public class LongJsonContentMiddlewareShould
 {
     private readonly LongJsonContentOptions _jsonContentOptions = new()
     {
-        MaxCharCountInField = 50,
+        MaxCharCountInField = 20,
         LeaveOnTrimCharCountInField = 10,
     };
 
-    private readonly IOptions<RequestLoggingOptions> _options;
+    private readonly IOptions<LongJsonContentOptions> _options;
 
-    public LongJsonContentMiddlewareTests()
+    public LongJsonContentMiddlewareShould()
     {
-        _options = Options.Create(new RequestLoggingOptions
-        {
-            LongJsonContent = _jsonContentOptions,
-        });
-
+        _options = Options.Create(_jsonContentOptions);
     }
 
     [Fact, Trait("Category", "Unit")]
     public void Constructor_OverridesDefaultsWithValuesFromConfig()
     {
-        // Arrange
         JsonStreamModifier jsonBuilder = new();
         _jsonContentOptions.MaxCharCountInField = 999;
         _jsonContentOptions.LeaveOnTrimCharCountInField = 888;
 
-        // Act
         var middleware = new LongJsonContentMiddleware(jsonBuilder, _options);
 
-        // Assert
         middleware.MaxCharCountInField.Should().Be(999);
         middleware.LeaveOnTrim.Should().Be(888);
     }
@@ -43,20 +38,18 @@ public class LongJsonContentMiddlewareTests
     [Fact, Trait("Category", "Unit")]
     public void Modify_ProperlyCreateComplex()
     {
-        // Arrange
+
         JsonStreamModifier jsonBuilder = new();
         _jsonContentOptions.MaxCharCountInField = 50;
         LongJsonContentMiddleware sut = new(jsonBuilder, _options);
-        var content = this.LoadResource("sis_vehicle_response.json");
+        string content = this.LoadResource("sis_vehicle_response.json");
         content = Regex.Replace(content, @"\s", "");
-        var bytes = Encoding.UTF8.GetBytes(content);
+        byte[] bytes = Encoding.UTF8.GetBytes(content);
         MemoryStream output = new();
 
-        // Act
         sut.Modify(new MemoryStream(bytes), output);
 
-        // Assert
-        var expected = this.LoadResource("sis_vehicle_response_log.json");
+        string expected = this.LoadResource("sis_vehicle_response_log.json");
         expected = Regex.Replace(expected, @"\s", "");
 
         new StreamReader(output).ReadToEnd()
@@ -67,17 +60,15 @@ public class LongJsonContentMiddlewareTests
     [Fact, Trait("Category", "Unit")]
     public void Modify_ProperlyCreateArray()
     {
-        // Arrange
+
         JsonStreamModifier jsonBuilder = new();
         LongJsonContentMiddleware sut = new(jsonBuilder, _options);
-        var content = "[\"1\",\"2\"]";
-        var bytes = Encoding.UTF8.GetBytes(content);
+        string content = "[\"1\",\"2\"]";
+        byte[] bytes = Encoding.UTF8.GetBytes(content);
         MemoryStream output = new();
 
-        // Act
         sut.Modify(new MemoryStream(bytes), output);
 
-        // Assert
         new StreamReader(output).ReadToEnd().Should()
             .NotBeEmpty()
             .And.BeEquivalentTo("[\"1\",\"2\"]");
@@ -86,17 +77,15 @@ public class LongJsonContentMiddlewareTests
     [Fact, Trait("Category", "Unit")]
     public void Modify_ProperlyCreateObject()
     {
-        // Arrange
+
         JsonStreamModifier jsonBuilder = new();
         LongJsonContentMiddleware sut = new(jsonBuilder, _options);
-        var content = "{\"1\":1,\"2\":true}";
-        var bytes = Encoding.UTF8.GetBytes(content);
+        string content = "{\"1\":1,\"2\":true}";
+        byte[] bytes = Encoding.UTF8.GetBytes(content);
         MemoryStream output = new();
 
-        // Act
         sut.Modify(new MemoryStream(bytes), output);
 
-        // Assert
         new StreamReader(output).ReadToEnd().Should()
             .NotBeEmpty()
             .And.BeEquivalentTo("{\"1\":1,\"2\":true}");
@@ -105,17 +94,15 @@ public class LongJsonContentMiddlewareTests
     [Fact, Trait("Category", "Unit")]
     public void Modify_ProperlyHandlesNullValue()
     {
-        // Arrange
+
         JsonStreamModifier jsonBuilder = new();
         LongJsonContentMiddleware sut = new(jsonBuilder, _options);
-        var content = "{\"1\":null,\"2\":true}";
-        var bytes = Encoding.UTF8.GetBytes(content);
+        string content = "{\"1\":null,\"2\":true}";
+        byte[] bytes = Encoding.UTF8.GetBytes(content);
         MemoryStream output = new();
 
-        // Act
         sut.Modify(new MemoryStream(bytes), output);
 
-        // Assert
         new StreamReader(output).ReadToEnd().Should()
             .NotBeEmpty()
             .And.BeEquivalentTo("{\"1\":null,\"2\":true}");
@@ -124,17 +111,15 @@ public class LongJsonContentMiddlewareTests
     [Fact, Trait("Category", "Unit")]
     public void Modify_ProperlyCreateNested()
     {
-        // Arrange
+
         JsonStreamModifier jsonBuilder = new();
         LongJsonContentMiddleware sut = new(jsonBuilder, _options);
-        var content = "{\"1\":1.1,\"object\":[\"1\",{\"2\":null}]}";
-        var bytes = Encoding.UTF8.GetBytes(content);
+        string content = "{\"1\":1.1,\"object\":[\"1\",{\"2\":null}]}";
+        byte[] bytes = Encoding.UTF8.GetBytes(content);
         MemoryStream output = new();
 
-        // Act
         sut.Modify(new MemoryStream(bytes), output);
 
-        // Assert
         new StreamReader(output).ReadToEnd().Should()
             .NotBeEmpty()
             .And.BeEquivalentTo("{\"1\":1.1,\"object\":[\"1\",{\"2\":null}]}");
@@ -143,17 +128,15 @@ public class LongJsonContentMiddlewareTests
     [Fact, Trait("Category", "Unit")]
     public void Modify_IgnoresNonJsonContentType()
     {
-        // Arrange
+
         JsonStreamModifier jsonBuilder = new();
         LongJsonContentMiddleware sut = new(jsonBuilder, _options);
-        var content = "This message is not a JSON string";
-        var bytes = Encoding.UTF8.GetBytes(content);
+        string content = "This message is not a JSON string";
+        byte[] bytes = Encoding.UTF8.GetBytes(content);
         MemoryStream output = new();
 
-        // Act
         sut.Modify(new MemoryStream(bytes), output);
 
-        // Assert
         new StreamReader(output).ReadToEnd().Should()
             .NotBeEmpty()
             .And.BeEquivalentTo("This message is not a JSON string");
@@ -162,17 +145,15 @@ public class LongJsonContentMiddlewareTests
     [Fact, Trait("Category", "Unit")]
     public void Modify_IgnoresNonJson()
     {
-        // Arrange
+
         JsonStreamModifier jsonBuilder = new();
         LongJsonContentMiddleware sut = new(jsonBuilder, _options);
-        var content = "This message is not a JSON string";
-        var bytes = Encoding.UTF8.GetBytes(content);
+        string content = "This message is not a JSON string";
+        byte[] bytes = Encoding.UTF8.GetBytes(content);
         MemoryStream output = new();
 
-        // Act
         sut.Modify(new MemoryStream(bytes), output);
 
-        // Assert
         new StreamReader(output).ReadToEnd().Should()
             .Be("This message is not a JSON string");
     }
@@ -180,96 +161,86 @@ public class LongJsonContentMiddlewareTests
     [Fact, Trait("Category", "Unit")]
     public void Modify_DoesNotModifySmallJson()
     {
-        // Arrange
+
         JsonStreamModifier jsonBuilder = new();
         _jsonContentOptions.MaxCharCountInField = 15;
         LongJsonContentMiddleware sut = new(jsonBuilder, _options);
-        var content = """{"key1":1,"key2":"some content"}""";
-        var bytes = Encoding.UTF8.GetBytes(content);
+        string content = @"{""key1"":1,""key2"":""some content""}";
+        byte[] bytes = Encoding.UTF8.GetBytes(content);
         MemoryStream output = new();
 
-        // Act
         sut.Modify(new MemoryStream(bytes), output);
 
-        // Assert
         new StreamReader(output).ReadToEnd().Should()
             .NotBeEmpty()
-            .And.BeEquivalentTo("""{"key1":1,"key2":"some content"}""");
+            .And.BeEquivalentTo(@"{""key1"":1,""key2"":""some content""}");
     }
 
     [Fact, Trait("Category", "Unit")]
     public void Modify_TrimsValue()
     {
-        // Arrange
+
         var jsonBuilder = new JsonStreamModifier();
         _jsonContentOptions.MaxCharCountInField = 10;
         LongJsonContentMiddleware sut = new(jsonBuilder, _options);
-        var content = """{"key1":"short","key2":"some long content"}""";
-        var bytes = Encoding.UTF8.GetBytes(content);
+        string content = @"{""key1"":""short"",""key2"":""some long content""}";
+        byte[] bytes = Encoding.UTF8.GetBytes(content);
         MemoryStream output = new();
 
-        // Act
         sut.Modify(new MemoryStream(bytes), output);
 
-        // Assert
         new StreamReader(output).ReadToEnd().Should()
             .NotBeEmpty()
-            .And.BeEquivalentTo("""{"key1":"short","key2":"some long ..."}""");
+            .And.BeEquivalentTo(@"{""key1"":""short"",""key2"":""some long ...""}");
     }
 
     [Fact, Trait("Category", "Unit")]
     public void Modify_TrimsCustomLengthValue()
     {
-        // Arrange
+
         JsonStreamModifier jsonBuilder = new();
         _jsonContentOptions.MaxCharCountInField = 10;
         _jsonContentOptions.LeaveOnTrimCharCountInField = 12;
+
         LongJsonContentMiddleware sut = new(jsonBuilder, _options);
 
-        var content = """{"key1":"short","key2":"some long content"}""";
-        var bytes = Encoding.UTF8.GetBytes(content);
+        string content = @"{""key1"":""short"",""key2"":""some long content""}";
+        byte[] bytes = Encoding.UTF8.GetBytes(content);
         MemoryStream output = new();
 
-        // Act
         sut.Modify(new MemoryStream(bytes), output);
 
-        // Assert
         new StreamReader(output).ReadToEnd().Should()
             .NotBeEmpty()
-            .And.BeEquivalentTo("""{"key1":"short","key2":"some long co..."}""");
+            .And.BeEquivalentTo(@"{""key1"":""short"",""key2"":""some long co...""}");
     }
 
     [Fact, Trait("Category", "Unit")]
     public void Modify_TrimsValueOnMultiDimension()
     {
-        // Arrange
+
         JsonStreamModifier jsonBuilder = new();
         _jsonContentOptions.MaxCharCountInField = 10;
         LongJsonContentMiddleware sut = new(jsonBuilder, _options);
-        var content = """{"key1":"short","key2":{"key1":"some long content"}}""";
-        var bytes = Encoding.UTF8.GetBytes(content);
+        string content = @"{""key1"":""short"",""key2"":{""key1"":""some long content""}}";
+        byte[] bytes = Encoding.UTF8.GetBytes(content);
         MemoryStream output = new();
 
-        // Act
         sut.Modify(new MemoryStream(bytes), output);
 
-        // Assert
         new StreamReader(output).ReadToEnd().Should()
             .NotBeEmpty()
-            .And.BeEquivalentTo("""{"key1":"short","key2":{"key1":"some long ..."}}""");
+            .And.BeEquivalentTo(@"{""key1"":""short"",""key2"":{""key1"":""some long ...""}}");
     }
 
     [Fact, Trait("Category", "Unit")]
     public void Modify_ProperlyFailsOnInvalidTrimValue()
     {
-        // Arrange
         JsonStreamModifier jsonBuilder = new();
 
-        // Act
         _jsonContentOptions.MaxCharCountInField = 0;
         Func<IRequestContentLogMiddleware> act = () => new LongJsonContentMiddleware(jsonBuilder, _options);
 
-        // Assert
         act.Should()
             .ThrowExactly<ArgumentOutOfRangeException>()
             .WithMessage("Specified argument was out of the range of valid values. (Parameter 'maxCharCountInField')");
